Reorg: Each day's solution will be a module
This commit restructures the project so that a single binary is produced
which imports each day as a module and expects two string arguments
(which day's solution to run and the string path to a file containing
its inputs).

Not generalized across solutions and in need of a better data structure
(look at those match statements— yuck!), plus some refactoring (see that
quick and dirty Path->String cast in each solution's main()).