Added a module "utils" to store snippets
The first snippet is for removing whitespace added by using multiline
string syntax. There are ways to hardcode string that don't add extra
whitespace, but they're not as a pretty. The added function is only
meant to be used in the tests (which have multiline inputs copied
directly from a webpage).